Move IKE constants and USIM/AKA helpers out of the emulator.
Keep IKE/ESP/TUN in swu_emulator.py; usim_aka.py and ikev2_const.py are structure-only extracts with the same behavior.
